+++ src/main/java/delivery_and_pickup_system/delivery_and_pickup_system/model/dto/pricingRule/PricingResponseDTO.java
@@ -3,8 +3,10 @@ package delivery_and_pickup_system.delivery_and_pickup_system.model.dto.pricingR
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PricingResponseDTO {
-    Double totalPrice;
+    BigDecimal totalPrice;
 }
